Make add/set/remove commands work from the server as well.

+++ src/main/java/com/azuredoom/levelingcore/commands/level/RemoveLevelCommand.java
@@ -1,29 +1,28 @@
 package com.azuredoom.levelingcore.commands.level;
 
-import com.hypixel.hytale.component.Ref;
-import com.hypixel.hytale.component.Store;
+import com.hypixel.hytale.server.core.Message;
 import com.hypixel.hytale.server.core.command.system.CommandContext;
 import com.hypixel.hytale.server.core.command.system.arguments.system.RequiredArg;
 import com.hypixel.hytale.server.core.command.system.arguments.types.ArgTypes;
-import com.hypixel.hytale.server.core.command.system.basecommands.AbstractPlayerCommand;
+import com.hypixel.hytale.server.core.command.system.basecommands.AbstractAsyncCommand;
 import com.hypixel.hytale.server.core.universe.PlayerRef;
-import com.hypixel.hytale.server.core.universe.world.World;
-import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
 import com.hypixel.hytale.server.core.util.Config;
 import com.hypixel.hytale.server.core.util.EventTitleUtil;
-import org.checkerframework.checker.nullness.compatqual.NonNullDecl;
+import org.jetbrains.annotations.NotNull;
 
+import java.util.concurrent.CompletableFuture;
 import javax.annotation.Nonnull;
 
 import com.azuredoom.levelingcore.LevelingCore;
 import com.azuredoom.levelingcore.config.GUIConfig;
 import com.azuredoom.levelingcore.lang.CommandLang;
+import com.azuredoom.levelingcore.utils.LevelingPlayerContextManager;
 
 /**
  * Represents a command that removes a specific number of levels from a player. This command operates within the
  * Leveling Core system and adjusts the player's level based on the specified number of levels to be removed.
  */
-public class RemoveLevelCommand extends AbstractPlayerCommand {
+public class RemoveLevelCommand extends AbstractAsyncCommand {
 
     @Nonnull
     private final RequiredArg<PlayerRef> playerArg;
@@ -45,22 +44,20 @@ public class RemoveLevelCommand extends AbstractPlayerCommand {
         this.levelArg = this.withRequiredArg("level", "Amount of levels to remove", ArgTypes.INTEGER);
     }
 
+    @NotNull
     @Override
-    protected void execute(
-        @NonNullDecl CommandContext commandContext,
-        @NonNullDecl Store<EntityStore> store,
-        @NonNullDecl Ref<EntityStore> ref,
-        @NonNullDecl PlayerRef playerRef,
-        @NonNullDecl World world
-    ) {
+    protected CompletableFuture<Void> executeAsync(@NotNull CommandContext commandContext) {
         var levelService = LevelingCore.getLevelService();
+
         if (levelService == null) {
             commandContext.sendMessage(CommandLang.NOT_INITIALIZED);
-            return;
+            return CompletableFuture.completedFuture(null);
         }
-        playerRef = this.playerArg.get(commandContext);
+
+        var playerRef = this.playerArg.get(commandContext);
         var levelRef = this.levelArg.get(commandContext);
         var playerUUID = playerRef.getUuid();
+
         var currentLevel = levelService.getLevel(playerUUID);
 
         if (currentLevel - levelRef <= 0) {
@@ -68,16 +65,56 @@ public class RemoveLevelCommand extends AbstractPlayerCommand {
                 CommandLang.CANNOT_REMOVE_LEVEL_BELOW_ONE
                     .param("player", playerRef.getUsername())
             );
-            return;
+            return CompletableFuture.completedFuture(null);
         }
-        levelService.removeLevel(playerUUID, levelRef);
-        var level = levelService.getLevel(playerUUID);
-        var removeLevelMsg = CommandLang.REMOVE_LEVEL_1.param("level", levelRef)
-            .param("player", playerRef.getUsername());
-        var levelTotalMsg = CommandLang.REMOVE_LEVEL_2.param("player", playerRef.getUsername()).param("level", level);
-        if (config.get().isEnableLevelAndXPTitles())
-            EventTitleUtil.showEventTitleToPlayer(playerRef, levelTotalMsg, removeLevelMsg, true);
-        commandContext.sendMessage(removeLevelMsg);
-        commandContext.sendMessage(levelTotalMsg);
+
+        var context = LevelingPlayerContextManager.getContext(playerUUID);
+
+        if (context == null || context.entityRef() == null || !context.entityRef().isValid()) {
+            commandContext.sendMessage(Message.raw("Could not find active player context."));
+            return CompletableFuture.completedFuture(null);
+        }
+
+        var future = new CompletableFuture<Void>();
+
+        context.world().execute(() -> {
+            try {
+                levelService.removeLevel(playerUUID, levelRef);
+
+                var level = levelService.getLevel(playerUUID);
+
+                var removeLevelMsg = CommandLang.REMOVE_LEVEL_1
+                    .param("level", levelRef)
+                    .param("player", playerRef.getUsername());
+
+                var levelTotalMsg = CommandLang.REMOVE_LEVEL_2
+                    .param("player", playerRef.getUsername())
+                    .param("level", level);
+
+                if (config.get().isEnableLevelAndXPTitles()) {
+                    EventTitleUtil.showEventTitleToPlayer(
+                        playerRef,
+                        levelTotalMsg,
+                        removeLevelMsg,
+                        true
+                    );
+                }
+
+                commandContext.sendMessage(removeLevelMsg);
+                commandContext.sendMessage(levelTotalMsg);
+
+                future.complete(null);
+            } catch (Exception e) {
+                LevelingCore.LOGGER.atWarning()
+                    .withCause(e)
+                    .log("Failed to remove level for player {}", playerUUID);
+
+                commandContext.sendMessage(Message.raw("Failed to remove level. Check the server log."));
+
+                future.complete(null);
+            }
+        });
+
+        return future;
     }
 }
