+++ src/main/java/com/azuredoom/levelingcore/commands/xp/AddXpCommand.java
@@ -1,24 +1,23 @@
 package com.azuredoom.levelingcore.commands.xp;
 
-import com.hypixel.hytale.component.Ref;
-import com.hypixel.hytale.component.Store;
+import com.hypixel.hytale.server.core.Message;
 import com.hypixel.hytale.server.core.command.system.CommandContext;
 import com.hypixel.hytale.server.core.command.system.arguments.system.RequiredArg;
 import com.hypixel.hytale.server.core.command.system.arguments.types.ArgTypes;
-import com.hypixel.hytale.server.core.command.system.basecommands.AbstractPlayerCommand;
+import com.hypixel.hytale.server.core.command.system.basecommands.AbstractAsyncCommand;
 import com.hypixel.hytale.server.core.universe.PlayerRef;
-import com.hypixel.hytale.server.core.universe.world.World;
-import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
 import com.hypixel.hytale.server.core.util.Config;
 import com.hypixel.hytale.server.core.util.EventTitleUtil;
-import org.checkerframework.checker.nullness.compatqual.NonNullDecl;
+import org.jetbrains.annotations.NotNull;
 
+import java.util.concurrent.CompletableFuture;
 import javax.annotation.Nonnull;
 
 import com.azuredoom.levelingcore.LevelingCore;
 import com.azuredoom.levelingcore.config.GUIConfig;
 import com.azuredoom.levelingcore.lang.CommandLang;
 import com.azuredoom.levelingcore.ui.hud.XPBarHud;
+import com.azuredoom.levelingcore.utils.LevelingPlayerContextManager;
 import com.azuredoom.levelingcore.utils.LevelingUtil;
 
 /**
@@ -27,7 +26,7 @@ import com.azuredoom.levelingcore.utils.LevelingUtil;
  * modification. It updates the player's XP and calculates the resulting level, sending feedback messages to both the
  * player and the command executor.
  */
-public class AddXpCommand extends AbstractPlayerCommand {
+public class AddXpCommand extends AbstractAsyncCommand {
 
     @Nonnull
     private final RequiredArg<PlayerRef> playerArg;
@@ -39,53 +38,99 @@ public class AddXpCommand extends AbstractPlayerCommand {
 
     public AddXpCommand(Config<GUIConfig> config) {
         super("addxp", "Add XP to player");
-        // this.requirePermission("levelingcore.addxp");
         this.config = config;
+
         this.playerArg = this.withRequiredArg(
             "player",
             "Player to add XP to.",
             ArgTypes.PLAYER_REF
         );
-        this.xpArg = this.withRequiredArg("xpvalue", "Amount of XP to add", ArgTypes.INTEGER);
+
+        this.xpArg = this.withRequiredArg(
+            "xpvalue",
+            "Amount of XP to add",
+            ArgTypes.INTEGER
+        );
     }
 
+    @NotNull
     @Override
-    protected void execute(
-        @NonNullDecl CommandContext commandContext,
-        @NonNullDecl Store<EntityStore> store,
-        @NonNullDecl Ref<EntityStore> ref,
-        @NonNullDecl PlayerRef playerRef,
-        @NonNullDecl World world
-    ) {
+    protected CompletableFuture<Void> executeAsync(@NotNull CommandContext commandContext) {
         var levelService = LevelingCore.getLevelService();
+
         if (levelService == null) {
             commandContext.sendMessage(CommandLang.NOT_INITIALIZED);
-            return;
+            return CompletableFuture.completedFuture(null);
         }
-        playerRef = this.playerArg.get(commandContext);
+
+        var playerRef = this.playerArg.get(commandContext);
         var xpRef = this.xpArg.get(commandContext);
         var playerUUID = playerRef.getUuid();
-        int maxLevel = LevelingUtil.computeMaxLevel();
-        int currentLevel = levelService.getLevel(playerUUID);
+
+        var maxLevel = LevelingUtil.computeMaxLevel();
+        var currentLevel = levelService.getLevel(playerUUID);
+
         if (currentLevel >= maxLevel) {
             commandContext.sendMessage(
                 CommandLang.ADD_LEVEL_MAX_LEVEL_REACHED
                     .param("player", playerRef.getUsername())
             );
-            return;
+            return CompletableFuture.completedFuture(null);
         }
-        levelService.addXp(playerUUID, xpRef);
-        XPBarHud.updateHud(playerRef);
-        int newLevel = levelService.getLevel(playerUUID);
-        if (newLevel > maxLevel) {
-            levelService.removeLevel(playerUUID, newLevel - maxLevel);
-            newLevel = maxLevel;
+
+        var context = LevelingPlayerContextManager.getContext(playerUUID);
+
+        if (context == null || context.entityRef() == null || !context.entityRef().isValid()) {
+            commandContext.sendMessage(Message.raw("Could not find active player context."));
+            return CompletableFuture.completedFuture(null);
         }
-        var setXPMsg = CommandLang.ADD_XP_1.param("xp", xpRef).param("player", playerRef.getUsername());
-        var levelTotalMsg = CommandLang.ADD_XP_2.param("player", playerRef.getUsername()).param("level", newLevel);
-        if (config.get().isEnableLevelAndXPTitles())
-            EventTitleUtil.showEventTitleToPlayer(playerRef, levelTotalMsg, setXPMsg, true);
-        commandContext.sendMessage(setXPMsg);
-        commandContext.sendMessage(levelTotalMsg);
+
+        var future = new CompletableFuture<Void>();
+
+        context.world().execute(() -> {
+            try {
+                levelService.addXp(playerUUID, xpRef);
+
+                XPBarHud.updateHud(playerRef);
+
+                var newLevel = levelService.getLevel(playerUUID);
+
+                if (newLevel > maxLevel) {
+                    levelService.removeLevel(playerUUID, newLevel - maxLevel);
+                    newLevel = maxLevel;
+                }
+
+                var setXPMsg = CommandLang.ADD_XP_1
+                    .param("xp", xpRef)
+                    .param("player", playerRef.getUsername());
+
+                var levelTotalMsg = CommandLang.ADD_XP_2
+                    .param("player", playerRef.getUsername())
+                    .param("level", newLevel);
+
+                if (config.get().isEnableLevelAndXPTitles()) {
+                    EventTitleUtil.showEventTitleToPlayer(
+                        playerRef,
+                        levelTotalMsg,
+                        setXPMsg,
+                        true
+                    );
+                }
+
+                commandContext.sendMessage(setXPMsg);
+                commandContext.sendMessage(levelTotalMsg);
+
+                future.complete(null);
+            } catch (Exception e) {
+                LevelingCore.LOGGER.atWarning()
+                    .withCause(e)
+                    .log("Failed to add XP for player {}", playerUUID);
+
+                commandContext.sendMessage(Message.raw("Failed to add XP. Check the server log."));
+                future.complete(null);
+            }
+        });
+
+        return future;
     }
 }
